Created new branch for SearchBar and committed changes to it
Had troubles getting the SearchBar functional.  Pretty sure it's my dev environment. Going to continue for the rest of the week forward, re-do the Beast over the weekend and get a fresh crack at it Monday
In the meantime, moving on in the main branch with other work
Going to commit to main the changes made for SearchBar branch, but discard most of the changes to get back to previous functionality and move forward with feature additions